[generator] inject_promo_cities is renamed to inject_promo_ids
