+++ tools/python/post_generation/inject_promo_ids.py
